Handle platforms where sizeof(funcptr) != sizeof(void*)
Fixes #4683

It's mostly a theoretical concern I think, but may apply to
some "not too obscure" platforms (ia64?).

For the majority of platforms it keeps working as before (copies
the function pointer into the void* of a capsule) although I've
used memcpy rather than a union to do this (because I don't
believe the union trick is 100% guaranteed to work in standard
c++, although practically it does of course).

For any where this doesn't apply it allocates memory instead and
the capsule points to that.

My interest in this is to use when pickling CyFunctions, but this
PR just uses the mechanism for cimporting cdef functions