tests/run/unicodemethods: Deduplicate tests where via-Cython output is compared to via-Python output
In every test, that is doing such comparison, replace the duplication
with a function that calls both py-function and cy-function, asserts the
result is the same and returns the result.

With this helper we can retain only one doctest entry instead of two for
every tested case. Tests also become a bit more robust because the
possibility of slight unintended error in between py and cy inputs is now
gone.

This refactoring is done in preparation to add more tests for unicode
methods that accept start and stop parameters.

Discussion: https://github.com/cython/cython/issues/4737#issuecomment-1100220378