Mark reverse operators with METHOD_COEXIST
This means that reverse operators (e.g. `__radd__`) won't be
hidden by the automatic wrapper that PyType_Ready produces if
the forward method exists. Although they won't work as
in Python, they will be possible to look up and call explicitly.
This should make it easier to write code that's compatible with
Cython 0.29.x and Cython 3 (where reverse operators will be
full supported).