Removed goto error from numpy cimport guard
The label it was using was unrelated to the place where the utility
code was actually appearing, and it was essentially luck that it
seemed to work (largly because it was always going to __pyx_L1_error
label).

It's guarded anyway because of
`writer.putln(writer.error_goto_if_PyErr(output.module_pos))`
which is added after every block of utility code in the init
section.

The only thing we lose is a position that links it back to the
"cimport numpy" line.