CI: Remove allow_failures for Python 3.11
Python 3.11 is now in beta and considered feature frozen. That means we can start helping making Python 3.11 more stable by fixing internal bugs and reporting upstream bugs.

This commit removes allowed_failure: true element of the CI configuration. It might happen that the CI will not always pass, but that only makes us aware of any existing and appearing bugs, and does not always have to be a strict blocker for PRs, especially in early beta.

Cython is one of the lowest level dependencies for many project, so early adoption of Python 3.11 can accelerate adoption across the whole ecosystem, especially as numpy and pandas are enabled to start to support Python 3.11.