Avoid unnecessary coercion to Python on int comparisons
Fixes some of https://github.com/cython/cython/issues/4819 (although
there's a second issue about applying this more generally to binops).

It works by always generating a second utility code function with
a bint result type and substituting that out if needed. This is
on the basis that the coercion is likely to be sufficiently
common that the code generation is worth the effort