Remove useless CppClassScope.inherited_type_entries attribute.
This attribute was mimicking the role of CppClassScope.inherited_var_entries,
but the situation is different: for inherited variables we use declare()
which doesn't store its entries in var_entries list, whereas for inherited
types we call declare_type(), which stores its entries in type_entries.
So, when inheriting types from a parent class, using base_scope.type_entries
is sufficient: we will also get the type entries of this base class' parents.