Move linetracing functions into an "exec"
Thus fixing linetracing tests of Python 3.11.

The issue we were seeing looked to be something to do with functions
defined within doctest docstrings. It was fixed by moving the
"defined-in-Python" functions into an exec call instead.

I'm not sure I quite understand the issue but it doesn't look
like a Python issue.