Handle extern PyVarObjects
Fixes https://github.com/cython/cython/issues/4827 at least
partially.

Some of patch copied from https://src.fedoraproject.org/rpms/Cython/pull-request/35#request_diff

Allows the size of a type to be between basicsize and basicsize+itemsize
since anything is this range is a reasonable size for a class to be.

Adds a test when an extern extension type is inherited from to
make sure that it isn't a PyVarObject.