avoid null dereference in __Pyx_KwValues_FASTCALL
Simpler follow up to https://github.com/cython/cython/pull/4726.
I don't think we need to be worried null args and non-zero nargs,
but null args and 0 nargs is quite common and valid I think.

This PR just avoids a dereference in that case (which i think is
probably dubious)