Make it easier to restore scanner state during parsing phase (GH-4813)
Things like match-case (essentially anything that uses Python's
new PEG parser capacities) are going to have to be implemented
by trying to parse something, failing, then going back and trying
to parse something else. This commit gets the initial work done
to make this easier to do.

Several error positions change in this effort, but this seems to improve the error reporting overall.