Remove "from x cimport class C"
(and "struct S"/"union U")

This appears to be an old Pyrex feature introduced in
https://github.com/cython/cython/commit/9bac9c2e014ed63c2ac435e9b431fae124fba668
to provide a way of "forward cimporting" a class. I originally tried to make
a test for it, but couldn't actually come up with a useful way
of using it in the intended way, where the name would be unavailable
initially but avaialble later.

It looks to be completely untested, and responsible for some missing
coverage in Nodes.py (https://github.com/cython/cython/issues/4163).
The large section containing
```
if kind == 'struct' or kind == 'union':  # 8479 ↛ 8480
```
I propose to fix the missing coverage by killing off the feature.

The only people I could find using this syntax were H5Py, who
look to have removed their sole use of it 3 years ago
https://github.com/h5py/h5py/commit/8d2498c7f5e3fec884ff56e9aca905c325d82484
Therefore it seems a good candidate to go in Cython 3