Implement the --depfile command-line option for the "cython" tool (GH-4916)
* refactor cythonize depfile support to be reusable

We would like to use it in the frontend for the cython tool as well.
Pull it out of the main loop and make it a utility function.

* fix depfiles emitting unusable relative paths

The paths for files in the source tree are relative instead of absolute.
This is not inherently a problem, but they are relative to the
containing directory of the source code file but the program that parses
these files (make, ninja) computes filenames relative to its working
directory, not relative to the source file.

Likewise, the final output file had its entire path component trimmed,
leaving just a bare filename.

Make this actually work by computing all relative paths relative to the
current working directory of the cython process itself. When invoked by
a build system, this will be the same directory the build system expects
files to be based on. This has the obvious drawback of requiring a sensible
work directory (i.e. you can't "just run" cython next to the sources), but it's
still a reasonable thing to assume and we already make that assumption in
other places.