Switch to YAML issue templates (#4971)
I think it's fairly hard to read the markdown forms and pick out instructional comments, bits that you're supposed to be modifying, etc. (even when you're trying to do it right and not just click through to submit your support question as quickly as possible).

Hopefully the YAML forms are a little clearer - they at least separate "instructions" from "user input" clearly.

I've also added a 3rd "other" free-form form for anything that doesn't quite fit into the two categories. We may regret that but it seemed sensible to have it as an option first.