Don't make SkipDeclarations a cdef class
It's only ever used as a mixin class. It ends up working because
it's empty (so doesn't cause a layout conflict) but that also means
it can't use any cdef features so there's no real benefit to
it being a cdef class. It also probably prevents some of the
transforms being cdef classes since Cython spots that it's the
second base.

See https://github.com/cython/cython/pull/4351 (which would
detect this misuse at runtime).