Remove `p_lambdef_nocond` from the parser, following Python 3.9+. (GH-4992)
Note that it wasn't correct before since it didn't pass the
correct flag to `p_lambdef` and thus was equivalent to just
using `p_lambdef`.

Note also that there's a difference in behaviour between Python3.9+
and before. Python <3.9 allowed `[i for i in range(10) if lambda: i]`
while Python >=3.9 disallows this. Arguably it's pointless because
the lambda always evaluates to True. See
https://github.com/python/cpython/issues/86014
for the Python issue.

With this change Cython will follow the Python 3.9 behaviour at
the cost of potentially breaking some code that does use the
pattern above.

Part of the cleanup in https://github.com/cython/cython/issues/4595