Improve detection of C complex.h
The existing check didn't really work (but I've left it...).
I think it's based on the quote:
> POSIX recommends checking if the macro _Imaginary_I is defined to
> identify imaginary number support.
My reading is that this is a way to determine if a special
imaginary type exists, but only after including the header. Until
you've included complex.h the macro won't be defined anyway.

From C11 there's a feature macro to determine if complex.h
exists so I've used that as the main way of detecting it.

In principle this could be backported to 0.29.x, but I don't
think it should be.