Fix mangling for .pxd cdef public functions (#5025)
This commit fixes a typo that caused `cdef public` functions from .pxd
files to be erroneously mangled. The unmangled name is chosen when the
definition from the .py file is processed, but on the condition that the
cname determined while processing the declaration from the .pxd file was
the mangled one (instead of a user-defined one).
The commit also adds a test for the ticket.