Restore 'int'->'object' for now
a44bbd363029aa9ba16fefcb485c68162f8ab663 changed an `int` annotation
so that it was interpreted as a Python `int` instead of a Python
object, which potentially breaks large integers on Python 2.

I think this was unintentional so I've added the special-cases
for this in. (I'm happy with this PR being rejected though if
it was intentional since I don't *really* care that much about
Python 2).

Fixes #4944