Fix builtin bound classmethods
The general optimization for unpacking builtin method calls
was breaking bound classmethods (with crashes on Py2 and wrong
answers on Py3).

This PR adds code to handle it. The code is unlikely to be
very fast, but my judgement is that this is a rare case and
thus it's more important just to get it to work rather than
to optimize this case.

Fixes #5051