Remove HTML template support from Tempita
HTML templates (and the associated functions/classes `sub_html`, `html`,
`attr`, and `html_quote`) were not used by Cython itself, but were left
over from the original, general-purpose, Tempita tool. The HTML
templating relied on the cgi package, which is deprecated as of Python
3.11 and will be removed with Python 3.13 (meaning that merely importing
the tempita package currently raises a DeprecationWarning). More
specifically, it relied on the function `cgi.escape`, which was already
removed with Python 3.8. This clearly suggests that the code has not
been used by anyone, since no-one seems to have complained about this
breakage.

Fixes #3309