Refactor module state to always use struct (#5056)
First part of https://github.com/cython/cython/issues/3689
First part of https://github.com/cython/cython/issues/4495

This tries to put all the global static variables into the
module state struct. When CYTHON_USE_MODULE_STATE is false,
they're looked up from a single static instance of the struct.
This allows a lot of semi-duplicated code to be dropped.

It doesn't get rid of the awkward `#defines` to access the
struct (that's probably the next step of the cleanup).

It moves the stringtab from a global into a function local.

It moves the statically defined coroutine types into the
module state struct in the same way that CyFunction/FusedFunction
already were.

It's mostly an exercise in deleting things - it doesn't
attempt to do any significant cleanup of the generation code.