Change how memoryviewslice/array has sequence flag set (#5187)
Registering cdef classes with collections.abs.sequence only works to set the type flag if cdef classes are mutable. Unfortunately we want them to be immutable since it's possible to corrupt memory with mutable cdef classes (#5023).

Therefore, add a Cython directive for marking cdef classes as a sequence, but only allow it in utility code for now. This is a cutdown version of #5030 - I don't want to add new public features without consensus, but I do want to merge the fix for mutable classes. Therefore I'm adding the feature privately for now.