Fix issues with partially optimized cascaded comparisons (GH-5357)
If a cascaded comparison is partially optimized (i.e. only some
of the comparisons are optimized) then the result types must end up
consistent all the way through. At the moment we select PyObject
which probably isn't the most efficient option, but is the
easiest to implement

We do not require the whole cascaded optimization to succeed.
Instead, we can just get Python comparisons as bool, and just ensure
that the entire cascade has the same type