Improve error diagnostic for exception-spec mismatch (#5547)
It seems like the most common Cython 3 snag is the change in
function exception behaviour. This PR adds some extra diagnostics
to help people navigate this change, and a general mechanism
for adding this kind of diagnostics.

It needs some tests, but I wasn't to see what tests fail
to tell me where to put them.