Fix issues with parallel/exception checked functions
Essentially the problem is that a `nogil` function just promises
that the function *can* be called without the GIL, not that
it actually doesn't have the GIL.

If you call `prange`/`parallel` inside that function while still
holding the GIL, if any thread has finished the GIL will be held
(... I think it's probably more of a mess than this, but mostly
you get away with it...). If a thread is still running and tries
to acquire the GIL then it'll deadlock permenantly.

I've solved this by adding an implicit `with nogil` around any
`prange`/`parallel` that's in a `nogil` block but isn't
certain that it has the GIL.

Fixes #5573 and #5564.