Allow 'cython -M' to work without distutils installed
When 'cython' is run specifying the -M command line option, the
Cython.Build.Dependencies module is imported. This modules also
contains support code to build Cython extensions via distutils or
setuptools. Importing it fails when running with Python 3.12, where
the distutils module has been removed.

Move one distutils import into the function that requires it. Handle
one import error and re-implement the simple utility function that is
required.

The Cython.Build module also re-exports the Cython.Distutils module.
To preserve the exported API, the only sensible thing is to allow the
corresponding import to fail when distutils is not present.

Fixes #5692.