Take CO_* from inspect in limited API
Closes #5619

This avoids "cheating" in the limited API by just redefining
stuff from the real API. Instead it looks it up at runtime instead
(which is probably still fragile, but a little more honest).

Also rewrites the loading of some of ModuleSetupCode to allow
bits of it to be defined in the impl/init sections, so not
everything has to be defined inline.