Fix C++ method overloading test when inheriting from an external class.
When overloading an inherited method, two cases arise: we either
inherit it from a class defined inside a cython code, or we inherit it
from a class wrapped in cython code but defined somewhere else
(an external C++ class).
In the latter case, the default exception behaviour is noexcept.
However, when overloading a c++ method in cython code, the default
exception behaviour is except?, which results in incompatible
signature types. This commit fixes that for the unit test
cpp_method_overloading.pyx by overloading inherited methods
from external classes by explicitely setting the exception behaviour
of the new methods to noexcept.