Pyupgrade: Update codebase to Python 3 syntax and features (GH-5810)
This commit updates the Cython code to Python 3 syntax and features, removing old, depreciated and redundant code. pyupgrade 3.15.0 with --keep-percent-format was used.

It only updates the code in the Cython directory, tests, docs and other directories aren't updated.

This includes:
- utf-8 encoding is now the default (PEP 3120, Python 3.0+)
- Replace list comprehensions by Generator Expressions (PEP 289, Python 2.4+)
- Replace yield loop by yield from (PEP 380, Python 3.3+)
- Replace the IOError alias by OSError (PEP 3151, Python 3.3+)
- Remove the default subclass (object) when defining a class
- Use New Super syntax (PEP 3135, Python 3.0+)
- Define sets with curly braces {} instead of set()
- Remove "r" parameter from open function, which is default
- Remove forced str("native") literals

The --keep-percent-format parameter ensures percent-formatted strings are not replaced with .format() strings. This could be done in another PR.