Fixes an address-sanitizer issue in parallel
Address sanitizers consistently gives an error on running the
"parallel" test, saying that it ends up getting freed
after releasing and reacquiring the gil, and thus we can't keep
using the same threadstate.

I'm not sure I understand threadstate well enough to know if this
is expected behaviour, and I suspect it pessimizes some of the
exception handling a little.