Add extra info on annotation inexact type (#5915)
For most builtin types, Cython interprets an annotation to mean
that an exact type is required. This is stricter than PEP-484 and
occassionally causes confusion.

1. Adds a bit more detail to the exception message in these cases
   (or to __notes__ in Python 3.12+) to explain.
2. Add a note in the migration guide.

I've only applied this to function arguments. Applying it to
general assignments would be possible but looks significantly more
complicated so I thought it better to cover the main case.

Related issue: https://github.com/cython/cython/issues/5908