Fix as_cfunction analysing args in wrong scope
Fixes #5970.

The basic cause of the issue was: if the argument in the .py file
has an annotation, then that annotation was analysed in the scope
of the .pxd file. Because the cimports were different in the two files
then it failed.

My view is that as_cfunction is called much too early to be reliably
analysing types in the .py file. Therefore, I've removed this.
Instead the arg type from the pxd file is stashed somewhere so that
it can be used later when the types from the .py file are properly
analysed. I think this is a worthwhile simplification.