Tighten the constness on some string arrays (#6032)
This seems to take about 1.5% off the size of
compiled "ParseTreeTransforms".

Essentially the original version is "an array of pointers to
unmodifiable strings". The new version is
"an unmodifiable array pointers to unmodifiable strings".
Presumably the advantage is that the compiler can know the
result at compile-time in a lot of cases.