Improve method inheritance in cppclass (#3235)
* Rewire calls to constructor and destructor in cppclass

* Give c++ methods entry the cfunction flag

* Fix c++ scope inheritance of func_cname

* Rework the symtab logic for C++ methods in Scope.declare

* Correct CppClassScope to allow method overloading with different signature

* Fix AttributeNode.analyse_as_type_attribute overloaded alternatives

* Allow c++ classes to have multiple constructors

* Amend error test cpp_no_constructor to reflect new logic in symtab for c++

* Fix test case cpp_classes_def: add expected warning

* Test case for cppclass method overloading

* Exclude cmethod entries from the overloadable functions

* Fix is_cfunction flag for inherited entries in cpp classes

* Guard is_cfunction flag setting in CppClassScope.declare_var

* cdef-extern-Docstring-ify the small Base class defined in C++ in cpp method overloading unit test

* When overloading an inherited method, override the entry inplace

* Update warning line for RegularPolygon.area method

* Fix comment wording.

Co-authored-by: da-woods <dw-git@d-woods.co.uk>

* Remove commented code

* Get rid of the void for default constructors of cppclass.

When generating the C++ code of a given cppclass, default constructor
can be generated. In these cases, it will be generated with and
explicit void argument. This commit gets rid of that void argument,
to look like what's generaly done by human.

* Move generate_cpp_constructor_code to class level.

generate_cpp_constructor_code is a helper function used in cppclass code generation.
Before this it captured some of its variables, and was nested inside the cppclass
code generation function. Now its separated, a proper method of ModuleNode,
being a function and not a closure.

* Comment the cppclass function alternatives overriding.

* Fix comment typo: c++ -> C++

* Commenting what's done when overriding a C++ method

* Update comment in Scope.declare() to match the code

The updated comment refers to what was `if type == alt_entry.type`
which is now `if type.compatible_signature_with(alt_entry.type)`.

* Fix transition from function to method of generate_cpp_constructor_code.

This is a fix about self argument and self.<function> call
instead of just <function>.

* Try to be clearer when overriding a method with a compatible signature.

* Clean C++ method overloading tests

* Harmonize numeric types for C++ method overloading tests

* Fix C++ method overloading test when inheriting from an external class.

When overloading an inherited method, two cases arise: we either
inherit it from a class defined inside a cython code, or we inherit it
from a class wrapped in cython code but defined somewhere else
(an external C++ class).
In the latter case, the default exception behaviour is noexcept.
However, when overloading a c++ method in cython code, the default
exception behaviour is except?, which results in incompatible
signature types. This commit fixes that for the unit test
cpp_method_overloading.pyx by overloading inherited methods
from external classes by explicitely setting the exception behaviour
of the new methods to noexcept.

* Fix typo in cpp_method_overloading.pyx comments

* Move logic for cppclass redeclarations outside declare

It's quite complex, and is complicating the base Scope declare
method. Instead put it in an overrideable method, so most of
the logic can be implemented by CppClassScope

* Other review suggestions

* Account of overriding C++ global names

Co-authored-by: gsamain <gwenael.samain@nexedi.com>