Reserve vector memory upfront (#6077)
Calling push_back in a loop caused it to perform repeated unnecessary memory allocations.

In this change, the necessary memory is reserved once ahead of the loop. This has a significant performance improvement for large data sets.

For objects which do not have a length, the old behaviour is preserved.