Make C++ conversion lengthhint limited-api friendly
Follow up to https://github.com/cython/cython/pull/6077.

I've had to make a change to allow Cython utility code to require
C utility code - I think that's worthwhile but there's other
ways rounds the problem if we don't want to do that (e.g.
just putting __Pyx_PyObject_LengthHint in the universal module
setup code).

I've picked the easy route and not implemented a limited API
version of length hint. It'd be perfectly possible to do
this, but it doesn't really seem like a priority.