[3.0.x] Fix c++ complex double_proj test
I think the issue was partly that the function was defined twice
with slightly different tests and that this was behaving differently
on Py2.7, but not completely sure.

Either way, it's testing the standard library rather than our
code, so we probably don't have to make it too detailed beyond
that it compiles.