Fix refcounting of default arguments with CYTHON_AVOID_BORROWED_REFS
This affects PyPy and Limited API I think.

References to the function default arguments were being incremented
but never decremeneted in the case where the argument was actually
provided. I suspect this mostly didn't matter, since arguments
are usually just globals with an extended lifetime, but it's
worth getting right.