Create standalone cython runtime module package
https://github.com/cython/cython/issues/1981

Package is named Cython_shadow. That can probably be changed.

It doesn't attempt to deal with any versioning issues.

It also doesn't do any testing. That'd probably be a case of
adding a ci run that makes and installs the wheel and then imports
a few example pure-Python modules.

Not sure what else I'm missing.