[3.0.x] Fix type of str call
If the type of a call to `str` has already been deduced as
unicode (because of language_level=3), then keep that type.

This was causing optimizations to fail, where the optimization
was made on the basis that the type was unicode and then that
was reversed by the replacement of `str`.

Fixes #6166.

Most of this doesn't need to go into 3.1.