Update documentation refcounting example
I think this makes some of the different behaviour clearer. The
initial example was somewhat complicated by things like
"is sys.getrefcount creating an extra reference?", and reading
global variables inside the functions which isn't
hugely relevant to much.

This tries to show the effect of creating new references inside
a function, which is more direct and probably shows what
we actually want.