Fix arg=str() default argument
Fixes #6192.

The basic issue is that the argument is deduced to be dynamic
default argument, and then at a late stage is optimized into
an empty string literal.

This means that correct handling code for the literal is skipped
because is has "is_dynamic" set, and the correct assignment code
into the dynamic struct is skipped because it's a literal.

I think the solution is to make the code that writes into the
dynamic struct more tolerant of literals.