Add thread-safety to some reference counting
Add a compiler directive to turn it off. Also turn it
off in the common case where a variable is completely
local to a function without a parallel block.

Implement DECREF_SET (and similar) using atomic swap so
that we know that the variable getting decrefed is the
same one that got swapped out. I've used the
(currently private) atomic operations built in to
Python since it seemed like the easiest option.

Also handle `del x`.

This is by no means complete (e.g. I think we'll have
to wrap reads in an atomic read too).