Fix use of __Pyx_TEST_large_func_pointers
Fixes #6236. It was defined in the main ModuleSetupCode which
meant it wasn't available in headers. Now it's no longer required
to be defined.

Also noted that __Pyx_capsule_to_c_func_ptr ends up in _api.h
files. Therefore gave it a version tag and an include guard so
that it's not required to be perpetually compatible.

I'm a little worried that the _api.h headers aren't tested well
enough and that we probably only use them from other Cython modules
rather than from an isolated C module.