Pre-clean docstrings when running EmbedSignature transform (#6241)
PEP257 documents rules for stripping inline docstrings such that they
may be formatted at the REPL using help.

Broadly, this is "strip whitespace from the first line, and dedent all
subsequent lines by the minimum common whitespace prefix".

This allows writing a docstring as:

```python
def function(*args) -> result:
    """This is the first line.

    Here is a long description.
    """
```

and having it rendered as:

```
This is the first line.

Here is a long description.
```

Similarly to if the docstring were:

```python
def function(*args) -> result:
    """
    This is the first line.

    Here is a long description.
    """
```

When compiling Cython files with embedsignature=True, the insertion of
the function signature breaks the normal cleaning rules, since the
raw docstring becomes:

```python
def function(*args) -> result:
    """function(*args) -> result
This is the first line.

    Here is a long description.
    """
```

Which is then rendered in the REPL as:

```
function(*args) -> result
This is the first line.

    Here is a long description.
```

Note the over-indented long description.

To fix this, change signature embedding to pre-clean the input
docstring (via inspect.cleandoc) before prepending the signature. The
above example is then embedded as:

```python
def function(*args) -> result:
    """function(*args) -> result
This is the first line.

Here is a long description.
"""
```
which is then rendered correctly.