Fix PyInteger / float division (#6195)
This was broken due to changes to type inference. It infers the
type as a double, but "py_operations" for binops need to have
a Python result. Therefore I've separated into a two step:

* calculate Python result
* coerce to known C type.

In principle Optimize.c could take this back to 1 step. But
at the moment it doesn't.

Fixes #6183
Fixes #6004