Optimize child_attrs property on non-CPython
This propery was a measurable issue on GraalPython (profiling
says ~17% of the total runtime, although replacing it only gets
9% speedup). PyPy also looks slightly faster with the
more straightforward Python code.

As far as I can tell, in CPython this operator.attrgetter ends
up in an optimized C code path so is probably worthwhile.
In most other implementations it ends up in a slightly slow
Python code-path that also loops over a list of attributes.

Since this property is looked up in TreeVisitor._visitchildren
it ends up quite significant to the performance.