Also avoid C++ keywords as identifiers.
We currently rename identifiers (such as struct fields) that
are C keywords; we should do the same for the larger set of
C++ keywords.

At this point in the code we do not have the context of
whether we're compiling in C++ or not, but it should be
safe to escape these (private) names a bit more agressively
even when it is not strictly necessary.