Narrow restrictions from !CYTHON_USE_TP_FINALIZE
Essentially as far as I can tell the only thing we can't really
do is PyObject_CallFinalizerFromDealloc.

So we can define `tp_finalize` in extension types. And that works
providing we don't define a custom deallocator (because the
default Python-supplied one calls it).

Also add a runtime warning when an extension type finalizer isn't
called.