Optimize refcounting in memoryviewslices
Closes https://github.com/cython/cython/issues/5507

Defines a set of rules where it's safe to make the temp a
borrowed reference when doing memoryview slicing. In that
case we can optimize for the reasonably likely case that
assignment needs no reference counting at all.

The two big things that this optimizes are:

```
for i in range(memview.shape[0]):
  subview = memview[i, :]
```

In this case only the first assignment is reference counted.

And

```
for i in range(memview.shape[0]):
  some_expression(memview[i, :])
```

where the whole thing is borrowed.