Get module-state working on multi-phase init
Essentially I've reimplemented the PyState_GetModule function
to provide a base-line fallback which will work (at least while
the GIL is held). Long-term I think we want to use the
proper mechanisms for isolating modules to get the state from
the class or the functions. However this achieves a base level
of functionality.

I've used binary-search in a list of thread ids. This is probably
needlessly inefficient (since practically the thread ids are
low sequential integers). However it should at least be fairly
general.