Pick a few small fixes out of #6226
So that they don't get lost based on being part of a larger
change that may not be our proposed behaviour.

A couple of `is_target` set on the lhs on generated assignments
that are generated too late to pick it up in analysis.

Skip `self.arg.generate_result_code` because it already gets
called indirectly via `ExprNode.generate_evaluation_code` ->
`ExprNode.generate_subexpr_evaluation_code`.