Disable some parallel tests in freethreading Python
Note that these tests are disabled because the current
behaviour doesn't work where more than one thread can be running
with the GIL.

i.e. to fix them would require a change in our intended behaviour
for parallel code rather than just a Cython code-generation fix.

(The intention here is to be able to turn off "allowed_failure"
on the freethreading branches)