Stop suggesting depending on `wheel` in PEP 518 config
`wheel` was never necessary, as old `setuptools` is auto-injecting it as a dep for wheel builds (but not sdist). And the modern one vendors it (`wheel` moved into `setuptools` during the PyCon 2024 sprint days).
`setuptools`' docs used to mistakenly imply that it was needed but I fixed those docs examples a few years ago. Let's correct this mistake in the Cython docs too.