Make generators thread safe (#6530)
Mostly for free-threading build but somewhere applicable anyway.

Improve thread safety of generators. The main mechanism to enforce that
was `is_running` so ensure that this is read and set pseudo-atomically.
This mostly means moving it a little further out in the implementation
(which is probably a good thing anyway).

Add some tests to try to stress-test the implementation at bit. They
definitely fail without the changes.