Add critical sections/thread safety to CyFunction accessors (#6582)
It's slightly unclear to me how useful this is - these are all
publically accessible read-write attributes that the user could
manipulate. However, there isn't a particularly good reason why they
would normally.

None of this should be on a performance-critical path though, and
cyfunctions are pretty ubiquitous everywhere, so it's probably
worthwhile.