Libcpp string reordering (#6632)
The first commit is the reordering according to the C++20 standard
draft. When comments are removed and lines are sorted, these are
semantic changes:

```diff
libcpp % diff -Naur string-orig.pxd string-new.pxd
--- string-orig.pxd	2025-01-25 14:18:57.670755478 +0100
+++ string-new.pxd	2025-01-25 14:26:23.365514610 +0100
@@ -135,11 +135,12 @@
 int stoi(const string& s, size_t* idx, int base) except +
 iterator begin()
 iterator end()
-iterator erase(const_iterator first, const_iterator last)
-iterator erase(const_iterator p)
-iterator erase(iterator first, iterator last)
-iterator erase(iterator p)
+iterator erase(const_iterator first, const_iterator last) except +
+iterator erase(const_iterator p) except +
+iterator erase(iterator first, iterator last) except +
+iterator erase(iterator p) except +
 iterator insert(iterator p, char c) except +
+iterator insert(iterator p, size_t n, char c) except +
 iterator insert(iterator, iterator, iterator) except +
 iterator operator+(size_type)
 iterator operator++()
@@ -279,7 +280,6 @@
 value_type& operator*()
 value_type& operator*()
 void clear()
-void insert(iterator p, size_t n, char c) except +
 void pop_back()
 void push_back(char c) except +
 void reserve(size_t) except +
 ```

These are missing `except +` declarations and the wrong return type for the `insert` method.

Second commit adds missing `replace' overload and adapts tests.

P.S. I didn't add the `const_iterator` versions to this header because, unlike `vector`, it gets really unmanageable, and it's always possible to do an explicit cast.