Stop Cython compiling itself in freethreaded builds (#6659)
Provide a mechanism where users can force it to if they want, but turn
it off by default.

---------------------------

As discussed in the comments of #6653, I think this is something that
shouldn't be enabled for most people. I don't want to stop anyone who
wants to, but making it off by default will stop pip doing it by
mistake.