Allow cimport of cpython in the limited API (#5871)
Fixes #5634.
Fixes #5697 (from a Cython point of view; Numpy still includes stuff
that isn't limited API compatible).

I've done two things:
1. Avoided a size check on extern types of `__builtins__` in the Limited API
(on the understanding that it probably won't be possible to use them).
2. Added a compiler directive for cdef functions that wraps them in an
`#if` guard. It's mainly for our internal use so undocumented but I'm
fairly relaxed if people find and use it.

With these I can get all of the pxd files (except cpython.array.pxd) so
that they can be be cimported in limited API mode. Of course, much of
them won't work, but it won't present a compile error unless used. Which
at least means that people can use the bits that do work.